Modify: separate Feed, Comment from Main.js
