+++ src/pages/jieunkim/Main/Main.scss
@@ -267,7 +267,7 @@ input:focus::placeholder {
 	border: 0;
 }
 
-span#author-id, span#comment-first-id, span#comment-second-id {
+span#author-id, span#comment-id {
 	font-weight: bold;
 }
